Blob of updates.
Modify Makefile to handle only sel4test
and sel4dynamic. Update logic.

Update Dockerfile. Cleanup, misc fixes
and logic improvements.

Handle all seL4 builds through build_sel4.sh.
Update/improve logic, improve param handling.

Update container enter logic.

Add minimal utils script.

+++ scripts/build_sel4.sh
@@ -1,38 +1,72 @@
-#! /bin/sh
+#!/usr/bin/env bash
 
-set -e
+set -ex
 
-# shellcheck disable=SC1091
-. "$(pwd)/.config"
+# Find utility functions
+SCRIPT_DIR="$(realpath $0)"
+SCRIPT_DIR="${SCRIPT_DIR%/*}"
+. "${SCRIPT_DIR}"/utils.sh
 
-# Detect whether should enter container:
-#   docker     -> '/.dockerenv' file exists
-#   lxc/podman -> 'container' variable is set to runtime name.
-#
-# With 'container' variable we support native builds too:
-# export container="skip", or similar, before calling make, and entering
-# container is skipped.
-if [ -z "${container}" ] && [ ! -f /.dockerenv ]; then
-  # shellcheck disable=SC2068
-  exec docker/enter_container.sh "$(pwd)" scripts/build_sel4.sh $@
+# Check number of args, if there are 4 we probably
+# aren't in container yet. Try to handle it.
+# Otherwise continue building.
+if [[ "$#" -eq 4 ]] \
+|| [[ "$#" -gt 4 ]]; then
+
+  # With 'container' variable we support native builds too:
+  # export container="skip", or similar, before calling make, 
+  # and entering container is skipped.
+  if [[ -z "${container}" ]] \
+  && [[ -z "${IN_CONTAINER}" ]]; then
+    # shellcheck disable=SC2068
+    exec docker/enter_container.sh "$1" "$2" "$3" "scripts/build_sel4.sh" "/workspace" "$4"
+  fi
 fi
 
-BUILDDIR=$1
-shift
+#build_sel4dynamic: $(CONFIG_FILE)
+#	@scripts/build_sel4dynamic.sh \
+#	$(WORKSPACE_ROOT) \
+#	$(CENGINE) \
+#	$(IMAGE):$(IMAGE_TAG) \
+#	projects/sel4_dynamic_loader
 
-SRCDIR=$1
 
-rm -rf "${BUILDDIR}"
+[[ "$#" -ne 2 ]] && die "Invalid # of arguments!" 
+
+WORKSPACE_ROOT="$(realpath $1)"
+[[ -z "${WORKSPACE_ROOT}" ]] && die "Invalid workspace root directory!"
+
+SRCDIR="$(realpath $2)"
+[[ -z "${SRCDIR}" ]] && die "Invalid source directory!"
+
+# Set suffix to the basename of the source directory.
+BUILDDIR_SUFFIX="${SRCDIR##*/}"
+
+# Get config
+# shellcheck disable=SC1091
+. "${WORKSPACE_ROOT}/.config"
+
+BUILDDIR="${WORKSPACE_ROOT}/build_${PLATFORM}_${BUILDDIR_SUFFIX}"
+[[ -z "${BUILDDIR}" ]] && die "Invalid build directory!"
+
+if [[ -e "${BUILDDIR}" ]] \
+&& [[ -d "${BUILDDIR}" ]]; then
+  rm -rf "${BUILDDIR}"
+fi
+
 mkdir -p "${BUILDDIR}"
 
-ln -rs tools/seL4/cmake-tool/init-build.sh "${BUILDDIR}"
+ln -rs "${WORKSPACE_ROOT}/tools/seL4/cmake-tool/init-build.sh" "${BUILDDIR}"
 ln -rs "${SRCDIR}/easy-settings.cmake" "${BUILDDIR}"
 
-cd "${BUILDDIR}" || exit 2
+pushd .
+cd "${BUILDDIR}" || die "Failed to enter build directory!"
 
 # shellcheck disable=SC2068
-./init-build.sh -B . -DAARCH64=1 -DPLATFORM="${PLATFORM}" -DCROSS_COMPILER_PREFIX="${CROSS_COMPILE}" $@
+#./init-build.sh -B . -DAARCH64=1 -DPLATFORM="${PLATFORM}" -DCROSS_COMPILER_PREFIX="${CROSS_COMPILE}" $@
+./init-build.sh -B . -DPLATFORM="${PLATFORM}" -DRELEASE=FALSE -DSIMULATION=TRUE
 ninja
+popd
 
-echo "Here are your binaries in ${BUILDDIR}/images: "
-ls -l ./images
+echo "Here are your binaries in ${BUILDDIR}: "
+ls -lA "${BUILDDIR}"/
